breaking build - split defs and component exports
